requirements.txt
autopep8==1.2.4
awscli==1.10.57
backports.ssl-match-hostname==3.5.0.1
boto==2.42.0
boto3==1.3.1
botocore==1.4.47
cached-property==1.3.0
colorama==0.3.7
docformatter==0.8
docker-compose==1.8.0
docker-py==1.9.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
enum34==1.1.6
falcon==1.0.0
functools32==3.2.3.post2
futures==3.0.5
gunicorn==19.6.0
httpie==0.9.6
ipaddress==1.0.16
jmespath==0.9.0
jsonschema==2.5.1
oauthlib==1.1.2
passlib==1.6.5
pbr==1.10.0
pep8==1.7.0
pep8radius==0.9.2
py4j==0.10.3
pyasn1==0.1.9
Pygments==2.1.3
python-dateutil==2.5.3
python-mimeparse==1.5.2
PyYAML==3.12
requests==2.7.0
requests-oauthlib==0.6.2
rsa==3.4.2
s3transfer==0.1.2
six==1.10.0
talons==0.3
texttable==0.8.4
untokenize==0.1.1
websocket-client==0.37.0
yapf==0.11.0
